requirements.txt
# Core Dependencies
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
idna==3.10
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyaes==1.6.1
pyasn1==0.6.1
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.0
requests==2.32.4
rsa==4.9.1
Telethon==1.40.0
urllib3==2.5.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
black>=23.0.0
